Editor: DataFileWriter lets to choose TextEncoder for writing game data
Unfortunately DataFileWriter is a fully static class, therefore we have to set the global DataFileWriter.TextEncoding whenever the game's text mode changes. This is of course not good, and should be fixed by e.g. refactoring DataFileWriter, splitting out value writing methods into a class that may be instantiated with particular Encoding.