Editor: added "GameTextEncoding" setting to the game
The property is a string ID of encoding. The TextEncodingTypeConverter class provides a list of encoding IDs. Currently it's a simple choice between ASCII/ANSI and UTF-8. Theoretically it may be expanded with explicit Windows-12xx codepages if necessary. Right now "ANSI" ID means "current system locale".

The setting is saved into the new OPT_GAMETEXTENCODING option in the compiled game as a codepage number.