Editor: in Scintilla disable preprocessor tracking for now
Scintilla supports tracking of #ifdef/#endif clauses, and separate styling for the disabled parts of the code.
However, it naturally cannot track pp definitions declared in different files (e.g. headers). For that to work we'd have to provide a list of definitions ourselves, similar to how we provide keyword list.