AGSNative: save text encoding hint in room files
Added a trivial optional "string map" extension to the room file. This may be used to store any kind of meta information.
For now, use this to add a text encoding hint, in case the room is imported from one game into another, or analyzed by some tool separately from the rest of the game.