Engine: support passing partial argument list to a script callback
In case the registered function has * lower * number of args, than provided by the caller (engine - in this case), skip the remaining, "unused", args instead of reporting an error.

This allows to run script callbacks with shorter argument list than the engine expects, and keep the script stack correct.
In other words, the callback arguments are treated as optional.

NOTE: in case the function has * higher * number of args, the engine will still report an error.