Engine: removed DrawAsSeparateCharSprite walk-behind method
This was an alternate walk-behinds method, probably meant for software renderer, but it's unclear if this was ever used on practice.
Judging by the algorithm it is slower than DrawOverCharSprite: it creates a cuts of the room background at the positions of characters and room objects, corresponding to the walk-behind, and draws them as separate sprites. This means that for each object there will be 2 sprites drawn (character/object + walk-behind cut) as opposed to 1 sprite draw (filtered character/object sprite) of DrawOverCharSprite. Also it was never fully optimized, as it draws a cut per *each* object, even if they are located at the same spot, thus may have overlapping cuts.