Engine: support Overlays not copying the image, but storing a sprite id
Overlays historically made a full copy of an assigned image. This was probably due to some limitations in dynamic sprite handling. But that may also lead to unnecessary waste of system resources.
Now we support storing sprite id refs in Overlay instead. For starters this will be only done when the assigned sprite is a normal (non-dynamic) sprite, as these cannot be deleted or edited at runtime.