Engine: fixed character's loop fixup in UpdateMoveAndAnim()
Was broken by 2ac1b258db203b62ba0613c979d7a092b51e441f

This old commit was trying to fix a problem in old game(s), when character was set to a loop with no frames. In such case the old engine would seek for the first loop with frames, starting with 0.
The condition for doing so was made incorrect though: instead of testing simply for a empty loop, it tested for frame number being outside of the current loop's frame range.
This broke games where e.g. some walking view's loops were shorter than the others.

A replacement is a condition that tests exactly for an empty loop instead. The frame exceeding a loop's frame count is fixed later along the way, so it's not a problem here.