Scintilla: CompareCaseInsensitive should use lowercase chars
This seem to be a widespread tradition to compare lowercase chars when doing case-insensitive comparison. At least this matches the .NET standard, which we use in the Editor.
The difference is that lowercase letters have values *higher* than the symbols (such as underscore '_' and so on), while uppercase have *lower* values than some of the symbols.

A use of uppercase letters in comparison breaks binary search in the autocomplete list, as it's sorted using "lowercase" rule. Changing to lowercase comparison instead, makes Scintilla match our Editor.
This fixes situations where presence of names with underscore ('_') character could cause autocomplete to malfunction.