Editor: use imported headers when creating autocomplete
Certain script elements require knowledge of the imported types. For example: extended (child) structs and extender functions. Without knowing which structs were declared in the other imported headers it may be impossible to properly register these.

This change makes it so that whenever AutoComplete.ConstructCache() is called, it also accepts a list of imported headers, from which it's possible to retrieve their own autocomplete cache. At the moment we only use a list of structs from these.