Engine: in OpenAlSource get PosMs by matching aloffset with buf records
Having a _positionMs member variable and using it to accumulate current playback position turned to be unreliable, as it has to be updated exactly when the OpenAL finishes processing a buffer; otherwise there are random situations when the retrieved AL_SEC_OFFSET does not correspond to the same timestamp as the last saved _positionMs. This primarily happens when looping a sound. But there may be other problems, such as when playback speed is changed: in which case al_offset may include the piece played with one speed and the piece played with another.

On the other hand, we already feature BufferRecords array, which tell timestamp and other info about currently queued al buffers. This array may be utilized when calculating current playback position; comparing precalculated buffer durations with al_offset we may find which buffers exactly have been passed, and in which buffer we're currently located, find its timestamp, and so on.