Common: fixed BufferedSectionStream not preventing read past end
There have been this stupid oversight all along since the introduction of the class: while it reported Length and EOS correctly, it still allowed to read past the end.
This can cause issues if the user code relies, for example, on Read() or ReadByte() return value.

In practice this fixes MP3 streaming, where the length of data is not always possible to precalculate.