Common: fixed BufferedSectionStream not preventing read past end
There have been this stupid oversight all along since the introduction of the class: while it reported Length and EOS correctly, it still allowed to read past the end.
This can cause issues if the user code relies, for example, on Read() or ReadByte() return value.

It is not known whether this was causing any trouble in practice as of the current engine version, but it could have.