Editor: fixed redundant operation could cause a crash in ViewComponent
The flatlist items swap was added as a part of 8c1d9b778cd8e2c83633a904d03576883a1f08f1, but it's not required for Views, as their item order in the flatlist is not used anyway, as they are ordered using their ID properties when written into compiled game.

At the same time this Swap operation is logically incorrect, as the flatlist may have been squashed after a View was deleted in the middle, while actual IDs would have gaps in the order. Changing ID operation may cause incorrect list assignment and even out of bounds exception in such case.