Engine: reimplemented Overlay.SetText without call to _display_main()
Previously Overlay.SetText was passing into CreateTextOverlay() -> Overlay_CreateTextCore() -> _display_main().
This effectively removed and recreated an overlay with different image but the same ID.

Also, previously, this method could have been only called for custom overlays in practice. For these the script object was not touched, and kept the correct overlay reference.

The situation has changed when we let user script to access some of the internally created overlays, such as Text and Portrait overlays created by a blocking Say command. When these overlays are recreated, they also must be completely destroyed, invalidating any existing script references. Hence calling _display_main() from Overlay.SetText() is no longer acceptable, as it leads to a full overlay recreation, and possible duplication of a script reference with two (or more) separate managed handles (this is almost like, two shared pointers owning same object).

To fix this problem, here we replace a call to CreateTextOverlay() with manual recreation of a textual image, and assigning a new image to the same overlay.
This does not change anything from the script's POV, as SetText method already suggests that the overlay reference remains valid (previously only the internal object was secretly replaced).