Engine: support optional size and position for gfxdriver's stage screen
This is primarily for backwards compatibility with the older plugins, that relied only on software drawing.

Previously we added "stage screens" to the hardware-accelerated graphic drivers (Direct3D/OpenGL), which provide a surface for plugins to draw upon. If used, these surfaces would then be rendered as plain sprites in 3D scene at certain place in the sprite sequence.

The remaining problem (left unnoticed) was that the surfaces which correspond to the two room render callbacks (right after background, and right after all objects) did not follow the room camera position and scaling correctly (not all, tbh). This commit addresses these problems by letting the engine to request certain size and optional position for these "stage screens" for particular sprite batches.