Improved snippet handling, 'for' loops
Provide snippet handling in cc_compiledscript.
- A 'Snippet' is a chunk of code that has been cut from the codebase.
- If a 'Restorepoint' is set and then some code is generated, then all bytes that have been generated since the 'Restorepoint' can be cut out and optionally moved into a 'Snippet'
- A 'Snippet' can 'Paste' itself to the current end of the codebase.

Introduce label handling, in cc_compiledscript
- If the value of a cell hasn't been determined yet, the cell can temorarily receive an integer called a 'Label'.
- Keep track of all cells that contain labels even when the piece of code that contains those cells is cut out and moved to a 'Snippet', or when the 'Snippet' is re-inserted into the code later on
- Keep a table that matches labels to values
- Provide a function that goes through the code and replaces labels by their values.

Rewrite 'yank' and 'write' in the parser to use 'Snippet' functions instead.  Kill 'yank' and 'write' in the parser.

Rewrite the callpoint and import handlers in the parser to use the 'Label' functionality instead. Kill the callpoint class in the parser.

Slightly improve tha hadling of 'for' loops, using the new 'Snippet' functionality.