Engine: in IAGSEngine::SetVirtualScreen set Stage buffer, not whole VS
When plugin calls SetVirtualScreen (for software renderer), don't replace whole virtual screen, instead replace only the current render stage buffer.
This is complementary to the older changes in software renderer, made during development of AGS 3.5.0, which featured advanced room viewports and cameras. Since that change, separate render stages could draw on sub-bitmaps of smaller size.

While IAGSEngine::GetVirtualScreen() was adjusted to follow that change, and return not the whole virtual screen, but only a "stage buffer" (which may be a VS sub-bitmap, or an intermediate bitmap created specifically for this render stage), SetVirtualScreen() was NOT adjusted accordingly and kept replacing whole VS. As a result, this discrepancy could cause logical errors, as well as crashes with plugins that use this API (e.g. original SnowRain plugin).

The immediate reason of error is that plugin would remember a pointer returned from GetVirtualScreen (which is a stage buffer), and then try to set it back with SetVirtualScreen. As a result, engine's own stage buffer is assigned as a full virtual screen.