Engine: support fixing up RoomStatuses restored from older save formats
This allows to properly fixup (upgrade) RoomStatus structs restored from a older format save in an upgraded game.
For example: game was upgraded from 3.4.0 to 3.6.0, but player tries loading 3.4.0 save. In this situation the engine must know and keep track of what version each RoomStatus was saved in.
This is necessary, because RoomStatus can only be updated upon loading a corresponding room file (when entering a room). This means that multiple RoomStatus objects will be staying in memory not being upgraded yet, until player enters particular room.