AGS.Native: write scripts using native methods again
The script writing was reimplemented in the managed code (C++ CLR) in c5c7c8748317ac7ae97c650ce1e6b2548aca2f54 , but this is quite inconvenient, as there's a script serialization code in the native part which is still used by the compiled room serialization code, and thus we'd have to duplicate any additions in native and managed code in parallel.

This change removes managed reimplementation, and makes CompiledScript::Write use native serialization instead.

Because the scripts are written in the middle of the compiled game format, which is serialized by the managed code too, we cannot use same stream object directly.
There are two options:
1) Write to a temp file using native stream, then copy file contents to the provided managed stream.
2) Write to a native membuf, then copy buf's contents to the managed stream (might require conversion, or marshaling).

I implemented this using a temp file for now. If this proves to be slow with large number of scripts, we may switch to using memory buffer (and memory stream) instead.