Editor: in OutputPanel sort compiler msg to have errors first
The purpose is user's convenience. The new compiler is now capable of writing multiple warnings, and if there are alot of them, the errors appear in the end of a very long list.

Added compiler message "index" number to keep the regular sort on. In theory they may also be sorted by the line number, but having an index seem more reliable. As there's also no numeric script module's ID, sorting by script's name would produce a different result.