Engine: restored --gfxfilter support for explicit scaling multipliers
The graphic mode config has now separate options for window size, game scaling and filter, but command line is lagging behind, so all the scaling is still set in `--gfxfilter`.
This is a hotfix for the time being, we shall think this over in the future, maybe introduce more args.