Engine: renamed WalkArea.Light to PlayerView, matching its real purpose
Apparently, this field was used to store light level of the walkable area in pre-2.55 engines, prior to introduction of the Regions.
In current engine, when loading games that old, this value is copied into Regions for backwards compatibility.
While in newer games it actually stores "player character view" override.

Renamed update_shadow_areas() into update_player_view(), and tidied the code for clarity.