Engine: a fix for the queued sounds (PlayQueue) starting with a gap
This is a slightly "hacky" way to fix the queued sounds starting with a small gap after the previous sound ends.
Because of the new audio playback subsystem in AGS 3.6.0, the clip start timing changed a little, and there appear to be small gap, about 1-2 game frame long, between previous and next sound in queue. The queue in AGS is not implemented ideally, because it relies on being updated once per game frame, rather than on an audio thread etc, but this is something we cannot change easily right now (maybe will reimplement whole thing later).

This commit does 2 things:
1. Sync logical channels with the audio subsystem also *prior* to updating queue and other things (crossfade etc). This fixes a 1 game frame queue delay.
2. Additionally, force queued clips to start 1 extra game frame earlier, by testing current playback position of active clips.