Engine: move cursor and cursor-over-gui updates to the game update fn
This is necessary for the touch-to-mouse emulation to work properly (and perhaps true touch controls in the future).
The order of updates is a bit wrong historically, where cursor position and gui control focus is updated later than the handling of button events. This means, for example, that if the cursor was positioned over button in game frame 1, then the mouse click will only trigger button press not later than the frame 2.

Changes:
* moved "cursor over gui" poll from draw_gui_and_overlays() to update_cursor_over_gui(), call it in the main game update fn.
* moved update_mouse_cursor() call from construct_game_screen_overlay() to the game update fn (renamed to update_cursor_view().
* picked out "cursor over location" trigger out of the "render" function into update_cursor_over_location(), call it explicitly in the main game update fn.

Effect on user script callbacks:
* the cursor position (and gui focus) will now be updated prior to `late_repeatedly_execute_always()` callback. Other callbacks will not be affected, as their relative order won't change compared with these updates.