Engine: when resetting objcache, invalidate spriteID, but not texture
Firstly, this is not necessarily optimal, as the next graphic assigned to these objects may be of same size, in which case the texture surface may be reused.

But there's another thing: there've been a bogus behavior in the AGS, where a dynamic sprite, deleted right before the room's exit transition, was still displayed for room objects (and everything else really).
This is because although the sprite was disposed, the image remained on a texture connected to these objects, at least until the next object update. But object updates don't run during room transitions, hence the effect.