Engine: removed alpha flags from DDB (textures)
NOTE: left opaque flag for now, as it makes bitmap->texture conversion faster for room bgs and few other things.
Still it's not designed very well.
Opaque should be either texture data's flag, - in which case same sprite_id will be either opaque or not opaque, - or DDB's flag, but in that case it cannot be applied to the shared texture data. Currently it's possible to share same texture data, but update it with different "opaque" values, which breaks logic.