Engine: ClearSharedDDB() should reset ID of the remaining texture data
This complements the original idea, but also as a side-effect fixes a mistake introduced by another hotfix 38d0dd3d7f0d6a4624d1f908f2e8d3c1abfee52a.
Because the ID remained intact in the texture, the engine would not know to stop using it and replace after a dynamic sprite was deleted.