Engine: added CCPluginObject, a thin wrapper over plugin dynobj manager
The purpose is to hide the actual plugin's manager behind a proxy, as plugin's interface may not fully comply to our internal one.
This prevents errors if one of the extended methods is called by the engine.
The base ICCDynamicObject interface currently consists of only few methods, which are used only once in the object's lifetime, and therefore this wrapper should have a small overhead.