Engine: ccInstance: cleaned up old commented code
This was an early attempt to pick out few operations into separate functions, but IIRC this caused additional slowdown. Maybe was not done in an optimal way.
In any case, it's better to remove the old commented code, as there may be slight differences between that and current inlined code now.
They may be recreated if necessary for another refactor. Just need to be aware of potential performance issues, and do tests.