Editor: implement base language in po translations
PO encoding will be ISO-8859-1 when ASCII for the most common case.
Though ASCII can be any 8bit encoding, in practice most people will find that common fonts like Arial are using a ISO8859-1 character set, so this should be the most compatible encoding to work with.

In the future we should likely enforce UTF8 always, once we have a solution for sprite fonts.