Editor: LogPanel: implemented toggleable "glue" mode, added AutoGlue button
When in "glue" mode the textbox is scrolling automatically along with the new entries.
The "glue" mode may be activated by user scrolling down to the end of the textbox. It is deactivated by user scrolling up.

The "autoglue" mode fixes texbox in a "glue" mode regardless of user's actions.

Note on implementation: we are using extension methods that pass SendMessage to the RichTextBox here, as C# RichTextBox class does not provide methods for manually setting scroll without changing selection.