Engine: brought multivariant script api functions to naming consistency
The idea is that the old variants have internal names with number (signifying number of args), while the latest variant has a name without number.

Make sure there's a distinct function for each variant, and older vars are calling later ones in a sequence. These calls likely to be inlined.