Engine: store ScriptUserObject in memory similar to DynamicArrays
I realized this recently that SUOs were not registered in the script exports properly. In AGS standards the object having a memory buffer should register that buffer's address directly, so that it could be read and written to directly. That is how strings, arrays, and any other structs are exported. This also how the engine plugins can access the data if they use script API to retrieve it.

There have not been any script API working with custom structs earlier, but there is since 3.5.0 (few functions returning managed Point struct). Plugins won't be able to use them correctly without this fix.