Engine: removed obscure hack from draw_button_background()
This removes a strange compatibility hack introduced by an ancient commit d638fdcd658e39e7a0158466213fe9d649d840ed
The hack was skipping drawing textwindow's background if it was a 1x1 transparent sprite. It's unclear what was the purpose for this, other than maybe a small drawing speed optimization?