SpriteCache: corrected handling of "remapped" sprites
Remapped sprites are the sprites that were supposed to be present in the game assets (according to the sprite index), but failed to load for whatever reasons.
The idea is that whenever a remapped sprite is requested, sprite cache should return a placeholder sprite (sprite 0) instead.

This commit corrects the logic of this, where it would use remapped sprite index one step earlier. This should fix incorrect duplicating addition of same "bad" sprite index to the MRU list, and also saves little redundant actions where the LoadSprite would reload the sprite 0 unnecessarily per each such request.