Engine: store game's ccInstances as unique_ptrs instead
The script handling part of the engine appeared to be too fragile and dependent on a precise timing of **deleting** a ccInstance.
For example, restoring a saved game may execute in the midst of the call to RunScriptFunction(), at the "post script cleanup" stage, while there are still references to ccInstance on the engine's call stack (and few other places). If anything prevents ccInstance's disposal at this moment, that would in turn lead to logical errors due to the game scripts "instances" counting. The most outstanding consequence is that the script exports are kept pointing into the deleted memory and not re-registered after the save is restored.

Overall, the script handling part of the engine would benefit from a major refactor, with a goal to make it safe and consistent, but this is not something I might do at the moment.