D3D/OGL: hotfix missing color depth record in Texture object
TODO:
DDB does not store a true texture's color depth for a while now, since we always use texture format corresponding to the display mode rather than the source bitmap format.
This needs to be adjusted, to avoid confusion in the future, but some thought has to be put into the DDB/Texture public interface, as color depth is currently used to test whether texture can be updated using given bitmap.