Partially revert "Common: made SpriteCache::GetFreeIndex() much faster"
This partially reverts commit 5cc9cfb6c537898ec6ef8181ec18fb8b04f9976f.

This change with queue of free IDs does not work with SpriteCache right away, because its interface has methods that set sprites to any arbitrary index (SetSprite). This may occur when restoring a saved game, in particular.
Either the interface should be changed, or another way to store free indexes (vector, hashmap?), or some trick to reset free indexes, but this has to be done in a safer way.

Also need to remember about Editor being able to SetSprite when changing sprite's ID, also to an arbitrary number, which case was also missed in this change.