Fix #1925 part 2
I had overlooked that by setting _editorTextModifiedSinceLastCopy with the current Modified value we should also manually update the Script. Because of this, going back in history and forward to the currently saved point would cause Script to keep a incorrect previous state.
With this commit ScriptEditor, Scintilla, and Script are synced whenever the user returns to the unmodified saved point.