Engine: a config hint to assume data version when reading a legacy save
I did a terribly stupid mistake when was expanding the old save format in the past (in pre-3.5.0 engines). Instead of checking for the save format, I've been checking for the game's data format.

This causes following problem: if a game made in a particular range of versions (something like 3.2.1 to 3.4.4) gets upgraded to a newer version (say 3.6.0), then all the old saves made in a pre-3.5.0 game version will not load correctly, as they are relying on the new compiled game data format.

Unfortunately there's no *proper* way to fix this, as the old save format had been in use for years, nothing can be done about it.
Hence the workaround that I may think of is to provide a config hint, that would tell the engine: whenever you load a legacy save format, assume that it was written by a game of particular data version.