Script API: implement Overlay.GetAtScreenXY, GetAtRoomXY, Visible
This adds:
* Overlay.Visible property;
* Overlay.GetAtScreenXY(x, y, room_overlays);
* Overlay.GetAtRoomXY(x, y);

Overlay.Visible property is simply an explicit way to turn overlays on and off rather than setting transparency to max. This makes overlays close to all the rest of the drawable objects in game in set of properties they use, and their logic. GetAtScreen/Room functions historically ignore the transparency and other visual effects, and only check for Enabled/Visible/Clickable properties (where applicable).
This property uses existing "flags" field, but had to increment save format number, because the "visible" flag needs to be set after loading previously made saves.

For GetAtScreen/Room functions I added a z-sorted vector of overlay indices, which is updated when:
* overlay is added,
* overlay is removed,
* overlay.ZOrder property changes
This vector is used to hit-test overlays in their z-order, as opposed to test them in the order of their storage in memory until the best match is found.
This should speed GetAt* functions up significantly in case of many overlays. However, I'm not certain if and how updating this vector will affect other actions, like adding/removing overlays.