Engine: changed ScreenOverlay to store owned images as dynamic sprites
This makes the storage and serialization more consistent, and also potentially allows to expose these images as dynamic sprite in script too, if will be desired in the future.
The exclusive images are no longer saved along with the overlay itself, but as a part of the dynamic sprite set. They will be still restored from older save format, but turned into dynamic sprites when resolving restored data.