Engine: smooth transition when ordering to walk while walking
This is an attempt to keep character moving seamlessly when a new walking order is made while the character is already walking.

Historically, engine had a problem with this, because of how a walking step is calculated. In short, a MoveList struct increments a number of "steps", but these "steps" may be less than pixel, in which case it takes >1 frames to make a visual change.
When a new walking order is made, this step counter is reset. This means that if player, for example, keeps clicking around, the character will be "stuttering".

In this commit we try following solution: if the character is already walking, then save last value of a fraction of "unfinished" pixel step, and apply it to the new movelist as a starting step counter.
This works pretty well for in case of manual orders.

I must note however, that still, due to the nature of this counter, there may be issues if a new Walk order is assigned each game frame. This may be done e.g. in script. In this case, depending on various factors (relation between x and y distance, walking speed, linking to animation, etc) the character may visually walk in a "ladder" instead of a straight line.
This is probably will not be noticed when commanded by a player, as players commonly do not click that frequently.

At this point I'm uncertain whether fixing the above would be only a matter of changing the types of variables (fixed -> float, in attempt to reduce conversion mistakes), or a matter of changing the core movelist mechanics.