Engine: restored faster String.Chars and Length variants for ASCII mode
The purpose is to at least keep string iteration relatively fast in ASCII game mode, for backwards compatibility.
The proper solution would be to use precalculated string length and use iterator, but that's not a task for 3.6.0 patches. See PR #2187 done for 3.6.1.

Also fixed String.Length for plugins in Unicode mode.