Editor: removed "force" arg from ScriptEditorBase::UpdateUICommands()
This parameter's meaning is not well defined and contradicts the purpose of the virtual method that may be overridden in the child classes.

Besides, it turns out that calling UpdateUICommands(force) from "GotFocus" event is not necessary, because UpdateUICommands is called on focus and other changes from Scintilla control's own event.