Editor: fixed RAM use skyrocketing during sprite reimport from sources
It appears that Editor's GC is not cleaning up temporary bitmaps in time, which may lead to temp data taking up all the memory, leading to program's failure.
This is easily seen when reimporting lots of tiles from a huge tileset image, for example.
Should force dispose everything temporary there.