Engine: unrestricted scriptname and name in CharacterInfo
NOTE 1: we *must* keep the old fixed-sized fields for the time being, for compaibility with the plugin API. Would the plugin API be adjusted and have direct access to a character data struct removed, then these fields may be removed as well.

NOTE 2: we still read these old fields from game data, avoid changing the data format in the middle (in the sake of both backward and forward compatibility). These fixed-sized values are assigned to the character properties first. For unrestricted values we'll have to add a data extension, appended to the end of data format.