Engine: unrestricted name in InventoryItemInfo
NOTE: we still read the old name field from game data, avoid changing the data format in the middle (in the sake of both backward and forward compatibility). These fixed-sized value is assigned to the item properties first. For unrestricted value we'll have to add a data extension, appended to the end of data format.