Engine: unrestricted script name an filename in ScriptAudioClip
NOTE: we still read the old fields from game data, avoid changing the data format in the middle (in the sake of both backward and forward compatibility). These fixed-sized values are assigned to the clip properties first. For unrestricted values we'll have to add a data extension, appended to the end of data format.