Engine: add unrestricted Character name to saves
NOTE: appended this new field to the end of character save chunk for consistency. But because the field itself is in the CharacterInfo struct, I have to pass CharacterInfo reference into CharacterExtra::ReadFromSavegame().
This is ugly, but but should be improved when the structs are refactored into having a distinct runtime Character class.