EXPERIMENT: up the OpenAlSource's MaxQueue
Our setting is done wrong, because a) it's constants regardless of situation, b) it counts only number of buffers instead of total length (or size).
First of all, buffers may be very small, and then if OpenAlSource is used by the system that does not care about it not being able to queue, then the playback will be skipping chunks.

The low queue limit was necessary mostly to workaround effects lag, such as setting speed. This may be resolved differently.