Engine: remade writing of non-fixed Character name in game saves
Done for more consistency, so long as CharacterInfo and CharacterExtras are separated structs.

WARNING: this breaks the saves made in 3.6.1.14 (Beta 15)...

Additionally, done so that legacy scriptnames are not loaded from current saves, since these are supposed to stay fixed throughout the game (at least unless characters are created dynamically, which they currently are not).