Editor: fix '\[' sequences lost in textual properties
The '[' is still supported as a linebreak in 3.* branch, for backwards compatibility. If user wants to have literal '[' they have to escape it like '\['.
But recent change 432ae84 have introduced Unescape call for properties (in order to support '\n' etc). Unescape call removes backslashes in unrecognized sequences.
So now we have to add this hack, escaping a backslash with another backslash, this will keep it intact.