Engine: support IsKeyPress for some of the NumPad keys
This only adds support for keys that do not depend on the NumLock state. This is because this function is not supposed to depend on modifiers state (same as IsKeyPressed in script).
This roughly corresponds to the pre-3.6.0 behavior.

OTOH, there is a distinct AGS KeyCode for NumPad5, but not others. Perhaps, it would be desired to add other NumPad* keycodes to let user combine them with System.NumLock mod to know the correct key state.