Engine: fixed CharacterInfo breaking script compatibility again
This fixes a mistake introduced in commit 4be1556, which in turn followed b02814b8bb3bb7a2cd00110466e2bd3047e772d1

The problem is, that CCStaticArray does not work as intended, and seems won't (at least not without some extra workarounds).

When the script compiler is told to get an Nth element of a global struct array, such as character[n], it calculates the memory address as
    Character* ptr = (array address + sizeof(Character) * n).
If this address is used for the read/write operations, these ops can be intercepted by interpreter and remapped into the real fields (see IScriptObject::ReadN, WriteN interface).
But if this address is used IN POINTER COMPARISON BY VALUE, then we cannot do anything. And if our real struct in the engine is stored on a different relative memory offset than one expected by compiler, then this pointer comparison will fail.
This breaks script expressions like e.g.:
      if (player == character[n])