Engine: fix backwards compatible Overlay z-sorting in creation order
Prior to introduction of explicit Overlay.ZOrder property, overlays were not sorted at all, and were drawn in the order of their storage in container, which coincidentally matched their order of creation.
After Overlay.ZOrder was added, and overlays are sorted among GUIs using qsort, there's a potential issue of their drawing order being undefined (if two overlays have matching or unassigned ZOrder).
This became a real issue after modification of the overlays internal storage, which was done for performance reasons.

The change introduces a global incrementing counter that assigns a "creation index" to Overlays, and is used as a second factor during sorting, similar to how GUIs use their ID property.